Criado Arquivo ErrosDict.py
Arquivo ErrorsDict para transformar a class errosdict em uma lib
