+++ smartcare-compose/apis/web-api/web-api.py
@@ -5,6 +5,7 @@ import os
 import psycopg2
 import requests
 from urllib.parse import parse_qs
+import ErrosDict as ErrosFile
 
 class PageError():
 
@@ -141,12 +142,12 @@ class UrlHandling():
             if repeats > 1:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(101)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(101)})
 
             elif repeats == 0:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(102)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(102)})
 
             else:
 
@@ -171,12 +172,12 @@ class UrlHandling():
                 except:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(102)+ " - Erro na segunda tentativa de encontrar a variável"})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(102)+ " - Erro na segunda tentativa de encontrar a variável"})
 
         except:
 
             Success = False
-            Errors.append({"msg": ErrorsDict.errorcode(103)})
+            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(103)})
 
         finally:
 
@@ -212,7 +213,7 @@ class UrlHandling():
         except:
 
             Success = False
-            Errors.append({"msg": ErrorsDict.errorcode(111)})
+            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(111)})
 
         finally:
 
@@ -246,12 +247,12 @@ class Env():
             else:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(201)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(201)})
 
         except:
 
             Success = False
-            Errors.append({"msg": ErrorsDict.errorcode(202)})
+            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(202)})
 
         finally:
 
@@ -283,7 +284,7 @@ class ConnectDataBase():
                     
                     WebApi.conn = None
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(301) + " - {0}".format(ex)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(301) + " - {0}".format(ex)})
 
             else:
 
@@ -297,7 +298,7 @@ class ConnectDataBase():
 
             Success = False
 
-            Errors.append({"msg": ErrorsDict.errorcode(302)})
+            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(302)})
 
         finally:
 
@@ -338,124 +339,12 @@ class ConnectDataBase():
         except:
 
             Success = False
-            Errors.append({"msg": ErrorsDict.errorcode(311)})
+            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(311)})
 
         finally:
 
             return json.dumps({"success": Success, "errors": Errors, "data": Data})
 
-class ErrorsDict():
-
-    def errorcode(int):
-
-        Errors = {}
-        #TRATAMENTO DE VARIAVEIS
-        Errors[101] = "A variável buscada foi passada mais de uma vez!"
-        Errors[102] = "A variável buscada não foi passada!"
-        Errors[103] = "Erro ao tratar as variáveis da requisição!"
-        Errors[104] = "O valor buscado deve ser diferente de vazio!"
-        Errors[105] = "Não foram passados parâmetros!"
-        Errors[106] = "Nenhuma coluna foi passada para a atualização!"
-        Errors[107] = "O id passado não pôde ser resolvido, ele deve ser um inteiro!"
-        Errors[108] = "Rota não encontrada!"
-        Errors[109] = "O valor buscado não é um inteiro!"
-        Errors[110] = "As variáveis esperadas não foram passadas ou estão incorretas!"
-        Errors[111] = "Erro extrair valores das variáveis do formulário!"
-        Errors[112] = "Variáveis obrigatórias não foram passadas!"
-        Errors[113] = "Os tipos das variáveis obrigatórias estão incorretos!"
-        Errors[114] = "Erro ao decodificar as variáveis!"
-        Errors[115] = "O tipo da variável está incorreto!"
-        #VARIAVEIS DE AMBIENTE
-        Errors[201] = "Não foi possivel resolver os dados de conexão da base de dados!"
-        Errors[202] = "Erro interno na api - Tradução de dados de conexão da base de dados!"
-        #CONEXÃO BASE DE DADOS
-        Errors[300] = "Erro na conexão com o banco de dados, contate o suporte!"
-        Errors[301] = "Erro ao se conectar à base de dados!"
-        Errors[302] = "Erro interno na Api - Conexão na Base de Dados!"
-        Errors[311] = "Erro ao autenticar a conexão com o banco de dados, contate o suporte!"
-        #TIPO DISPOSITIVO
-        Errors[401] = "Erro na listagem de tipo de dispositivo!"
-        Errors[411] = "Erro na busca de tipos!"
-        Errors[412] = "Erro interno na Api - busca tipo por id!"
-        Errors[421] = "Erro interno na Api - busca tipo por nome!"
-        Errors[431] = "A inserção foi bem sucedida, porém não encontramos os dados do tipo no banco!"
-        Errors[432] = "Erro na inserção do tipo!"
-        Errors[433] = "Erro interno na Api - inserir tipo!"
-        Errors[441] = "A atualização do tipo foi bem sucedida, porém não encontramos os dados do tipo no banco!"
-        Errors[442] = "Erro na atualização do tipo!"
-        Errors[443] = "O id do tipo de dispositivo não foi localizado!"
-        Errors[444] = "Erro interno na Api - atualizar tipo!"
-        Errors[451] = "O nome do tipo a ser alterado não foi passado ou está incorreto!"
-        Errors[452] = "Erro interno na Api - Atualizar nome do tipo!"
-        Errors[461] = "A exclusão do tipo não foi bem sucedida!"
-        Errors[462] = "A exclusão foi bem sucedida, mas ocorreu um erro ao buscar os índices do tipo!"
-        Errors[463] = "Erro na exclusão do tipo!"
-        Errors[464] = "Tipo já excluído!"
-        Errors[465] = "Erro interno na Api - excluir tipo!"
-        #500 - DISPOSITIVO
-        Errors[501] = "Erro na listagem de dispositivo!"
-        Errors[511] = "Erro interno na Api - busca dispositivo por id!"
-        Errors[521] = "Erro na busca de dispositivo!"
-        Errors[522] = "Erro interno na Api - busca dispositivo por texto!"
-        Errors[531] = "A inserção foi bem sucedida, porém não encontramos os dados do dispositivo no banco!"
-        Errors[532] = "Erro na inserção dos dispositivo!"
-        Errors[533] = "Erro interno na Api - inserir dispositivo!"
-        Errors[541] = "A atualização foi bem sucedida, porém não encontramos os dados do dispositivo no banco!"
-        Errors[542] = "Erro na atualização do dispositivo!"
-        Errors[543] = "O id do dispositivo não foi localizado!"
-        Errors[544] = "Erro interno na Api - atualizar dispositivo!"
-        Errors[551] = "O nome do dispositivo a ser alterado não foi passado ou está incorreto!"
-        Errors[552] = "Erro interno na Api - Atualizar nome do dispositivo!"
-        Errors[561] = "O codigo do dispositivo a ser alterado não foi passado ou está incorreto!"
-        Errors[562] = "Erro interno na Api - Atualizar codigo do dispositivo!"
-        Errors[571] = "A exclusão do dispositivo não foi bem sucedida!"
-        Errors[572] = "A exclusão foi bem sucedida, mas ocorreu um erro ao buscar os índices do dispositivo!"
-        Errors[573] = "Erro na exclusão do dispositivo!"
-        Errors[574] = "Dispositivo já excluído!"
-        Errors[575] = "Erro interno na Api - excluir dispositivo!"
-        #600 - AMBIENTE
-        Errors[601] = "Erro na listagem de ambiente!"
-        Errors[611] = "Erro na busca de ambiente!"
-        Errors[612] = "Erro interno na Api - busca ambiente por id!"
-        Errors[621] = "Erro interno na Api - busca ambiente por texto!"
-        Errors[631] = "A inserção foi bem sucedida, porém não encontramos os dados do ambiente no banco!"
-        Errors[632] = "Erro na inserção dos ambiente!"
-        Errors[633] = "Erro interno na Api - inserir ambiente!"
-        Errors[641] = "A atualização foi bem sucedida, porém não encontramos os dados do ambiente no banco!"
-        Errors[642] = "Erro na atualização do ambiente!"
-        Errors[644] = "O id do ambiente não foi localizado!"
-        Errors[645] = "Erro interno na Api - atualizar ambiente!"
-        Errors[651] = "O nome do ambiente a ser alterado não foi passado ou está incorreto!"
-        Errors[652] = "Erro interno na Api - Atualizar nome do ambiente!"
-        Errors[661] = "Erro ao atualizar a descrição do ambiente!"
-        Errors[662] = "A descrição do ambiente a ser alterado não foi passado ou está incorreto!"
-        Errors[663] = "Erro interno na Api - Atualizar descrição do ambiente!"
-        Errors[671] = "A exclusão do ambiente não foi bem sucedida!"
-        Errors[672] = "A exclusão foi bem sucedida, mas ocorreu um erro ao buscar os índices do ambiente!"
-        Errors[673] = "Erro na exclusão do ambiente!"
-        Errors[674] = "Ambiente já excluído!"
-        Errors[675] = "Erro interno na Api - excluir ambiente!"
-        #700 - MEDICAO
-        Errors[701] = "As datas passadas estão num formato inválido!"
-        Errors[702] = "Erro ao Buscar medições!"
-        #800 - USUARIO
-        Errors[801] = "Login não encontrado!"
-        Errors[802] = "Erro ao efetuar Login!"
-        Errors[803] = "Erro ao tratar Login e Senha!"
-        Errors[811] = "A inserção foi bem sucedida, porém não encontramos os dados do usuário no banco!"
-        Errors[812] = "Erro na inserção do usuário!"
-        Errors[813] = "Senha inválida! Ela deve ter mais de 6 dígitos e deve conferir com a confirmação!"
-        Errors[814] = "Erro interno na Api - inserir usuário!"
-
-        error = Errors[int] if int in Errors.keys() else None
-
-        if error != None:
-
-            return error
-
-        else:
-
-            return f"Ocorreu um erro, entre em contato com o suporte! (error code {int})"
 
 class WebApi(Bottle):
 
@@ -694,7 +583,7 @@ class WebApi(Bottle):
 
                 self.conn.rollback()
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(401)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(401)})
 
             finally:
 
@@ -705,7 +594,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -751,7 +640,7 @@ class WebApi(Bottle):
 
                             self.conn.rollback()
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(411)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(411)})
 
                         finally:
 
@@ -760,12 +649,12 @@ class WebApi(Bottle):
                     elif str(idbusca) == "":
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
 
                     else:
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(109)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(109)})
 
                 else:
 
@@ -778,7 +667,7 @@ class WebApi(Bottle):
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(412)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(412)})
 
             finally:
 
@@ -787,7 +676,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -835,7 +724,7 @@ class WebApi(Bottle):
 
                             self.conn.rollback()
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(411)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(411)})
 
                         finally:
 
@@ -844,7 +733,7 @@ class WebApi(Bottle):
                     else:
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
                 else:
 
                     Success = False
@@ -856,7 +745,7 @@ class WebApi(Bottle):
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(421)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(421)})
 
             finally:
 
@@ -865,7 +754,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -908,13 +797,13 @@ class WebApi(Bottle):
                         if not Data:
 
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(431)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(431)})
                     
                     except:
 
                         self.conn.rollback()                    
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(432)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(432)})
                     
                     finally:
                     
@@ -923,12 +812,12 @@ class WebApi(Bottle):
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(110)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(110)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(433)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(433)})
             
             finally:
             
@@ -937,7 +826,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -962,7 +851,7 @@ class WebApi(Bottle):
                 if not FormData:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(105)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(105)})
 
                 elif idtipo != None and str(idtipo).isnumeric():
 
@@ -1006,13 +895,13 @@ class WebApi(Bottle):
                                 if not Data:
                                 
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(441)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(441)})
                             
                             except:
 
                                 self.conn.rollback()                            
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(442)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(442)})
                             
                             finally:
                             
@@ -1021,12 +910,12 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(106)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(106)})
                     
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(443)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(443)})
 
                         for error in DataBeforeErrors:
 
@@ -1035,12 +924,12 @@ class WebApi(Bottle):
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(444)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(444)})
             
             finally:
             
@@ -1049,7 +938,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -1105,13 +994,13 @@ class WebApi(Bottle):
                                 if not Data:
 
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(441)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(441)})
                             
                             except:
 
                                 self.conn.rollback()                            
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(442)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(442)})
                             
                             finally:
                             
@@ -1120,7 +1009,7 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(443)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(443)})
 
                             for error in DataBeforeErrors:
 
@@ -1129,17 +1018,17 @@ class WebApi(Bottle):
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(451)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(451)})
                 
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(452)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(452)})
             
             finally:
             
@@ -1148,7 +1037,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -1224,13 +1113,13 @@ class WebApi(Bottle):
                                     if Data:
 
                                         Success = False
-                                        Errors.append({"msg": ErrorsDict.errorcode(461)})
+                                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(461)})
 
                                 except:
 
                                     self.conn.rollback()
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(462)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(462)})
 
                                 finally:
                                 
@@ -1240,7 +1129,7 @@ class WebApi(Bottle):
 
                                 self.conn.rollback()                
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(463)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(463)})
                             
                             finally:
 
@@ -1249,12 +1138,12 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(464)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(464)})
                     
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(443)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(443)})
 
                         for error in DataBeforeErrors:
 
@@ -1263,12 +1152,12 @@ class WebApi(Bottle):
                 else:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
 
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(465)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(465)})
 
             finally:
 
@@ -1277,7 +1166,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -1325,7 +1214,7 @@ class WebApi(Bottle):
 
                 self.conn.rollback()
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(501)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(501)})
             
             finally:
             
@@ -1335,7 +1224,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -1391,7 +1280,7 @@ class WebApi(Bottle):
                         else:
 
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(108)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(108)})
 
                             return({"success":Success,"errors":Errors,"data":Data})
                 
@@ -1418,7 +1307,7 @@ class WebApi(Bottle):
 
                             self.conn.rollback()                           
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(501)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(501)})
                         
                         finally:
                         
@@ -1427,12 +1316,12 @@ class WebApi(Bottle):
                     elif str(idbusca) == "":
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
                     
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(109)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(109)})
                 
                 else:
                 
@@ -1446,7 +1335,7 @@ class WebApi(Bottle):
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(511)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(511)})
             
             finally:
             
@@ -1455,7 +1344,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -1499,7 +1388,7 @@ class WebApi(Bottle):
                         else:
 
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(108)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(108)})
 
                             return({"success":Success,"errors":Errors,"data":Data})
 
@@ -1521,7 +1410,7 @@ class WebApi(Bottle):
             
                             self.conn.rollback()                            
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(521)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(521)})
                         
                         finally:
                         
@@ -1530,7 +1419,7 @@ class WebApi(Bottle):
                     else:
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
                 
                 else:
                 
@@ -1544,7 +1433,7 @@ class WebApi(Bottle):
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(522)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(522)})
             
             finally:
             
@@ -1553,7 +1442,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -1659,13 +1548,13 @@ class WebApi(Bottle):
                             if not Data:
 
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(531)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(531)})
                         
                         except:
                             
                             self.conn.rollback()
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(532)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(532)})
                         
                         finally:
                         
@@ -1674,22 +1563,22 @@ class WebApi(Bottle):
                     except:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(533)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(533)})
 
                 elif MandatoryVarsExists == False:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(112)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(112)})
 
                 else: #MandatoryVarsTypes == False
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(113)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(113)})
             
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(114)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(114)})
 
             finally:
 
@@ -1698,7 +1587,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -1723,7 +1612,7 @@ class WebApi(Bottle):
                 if not FormData:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(105)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(105)})
 
                 elif iddispositivo != None and str(iddispositivo).isnumeric():
 
@@ -1830,13 +1719,13 @@ class WebApi(Bottle):
                                     if not Data:
                                     
                                         Success = False
-                                        Errors.append({"msg": ErrorsDict.errorcode(541)})
+                                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(541)})
                                 
                                 except:
 
                                     self.conn.rollback()                                
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(542)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(542)})
                                 
                                 finally:
                                 
@@ -1845,27 +1734,27 @@ class WebApi(Bottle):
                             else:
                             
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(106)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(106)})
 
                         else:
 
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(115)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(115)})
                         
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(543)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(543)})
                     
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(544)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(544)})
             
             finally:
             
@@ -1874,7 +1763,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -1930,13 +1819,13 @@ class WebApi(Bottle):
                                 if not Data:
 
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(541)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(541)})
                             
                             except:
 
                                 self.conn.rollback()                            
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(542)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(542)})
                             
                             finally:
                             
@@ -1945,7 +1834,7 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(543)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(543)})
 
                             for error in DataBeforeErrors:
 
@@ -1954,17 +1843,17 @@ class WebApi(Bottle):
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(551)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(551)})
                 
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(552)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(552)})
             
             finally:
             
@@ -1973,7 +1862,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2029,13 +1918,13 @@ class WebApi(Bottle):
                                 if not Data:
 
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(541)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(541)})
                             
                             except:
 
                                 self.conn.rollback()                            
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(542)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(542)})
                             
                             finally:
                             
@@ -2044,7 +1933,7 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(543)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(543)})
 
                             for error in DataBeforeErrors:
 
@@ -2053,17 +1942,17 @@ class WebApi(Bottle):
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(561)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(561)})
                 
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(562)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(562)})
             
             finally:
             
@@ -2072,7 +1961,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2148,13 +2037,13 @@ class WebApi(Bottle):
                                     if Data:
 
                                         Success = False
-                                        Errors.append({"msg": ErrorsDict.errorcode(571)})
+                                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(571)})
 
                                 except:
 
                                     self.conn.rollback()
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(572)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(572)})
 
                                 finally:
                                 
@@ -2164,7 +2053,7 @@ class WebApi(Bottle):
 
                                 self.conn.rollback()                
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(573)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(573)})
                             
                             finally:
 
@@ -2173,12 +2062,12 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(574)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(574)})
                     
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(543)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(543)})
 
                         for error in DataBeforeErrors:
 
@@ -2187,12 +2076,12 @@ class WebApi(Bottle):
                 else:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
 
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(575)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(575)})
 
             finally:
 
@@ -2201,7 +2090,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2245,7 +2134,7 @@ class WebApi(Bottle):
 
                 self.conn.rollback()
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(601)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(601)})
             
             finally:
             
@@ -2255,7 +2144,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2301,7 +2190,7 @@ class WebApi(Bottle):
                         else:
 
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(108)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(108)})
 
                             return({"success":Success,"errors":Errors,"data":Data})
                 
@@ -2328,7 +2217,7 @@ class WebApi(Bottle):
 
                             self.conn.rollback()                           
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(611)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(611)})
                         
                         finally:
                         
@@ -2337,12 +2226,12 @@ class WebApi(Bottle):
                     elif str(idbusca) == "":
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
                     
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(109)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(109)})
                 
                 else:
                 
@@ -2356,7 +2245,7 @@ class WebApi(Bottle):
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(612)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(612)})
             
             finally:
             
@@ -2365,7 +2254,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2409,7 +2298,7 @@ class WebApi(Bottle):
                         else:
 
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(108)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(108)})
 
                             return({"success":Success,"errors":Errors,"data":Data})
 
@@ -2431,7 +2320,7 @@ class WebApi(Bottle):
             
                             self.conn.rollback()                            
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(611)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(611)})
                         
                         finally:
                         
@@ -2440,7 +2329,7 @@ class WebApi(Bottle):
                     else:
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
                 
                 else:
                 
@@ -2454,7 +2343,7 @@ class WebApi(Bottle):
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(621)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(621)})
             
             finally:
             
@@ -2463,7 +2352,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2531,13 +2420,13 @@ class WebApi(Bottle):
                             if not Data:
 
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(631)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(631)})
                         
                         except:
                             
                             self.conn.rollback()
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(632)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(632)})
                         
                         finally:
                         
@@ -2546,22 +2435,22 @@ class WebApi(Bottle):
                     except:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(633)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(633)})
 
                 elif MandatoryVarsExists == False:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(112)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(112)})
 
                 else: #MandatoryVarsTypes == False
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(113)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(113)})
             
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(114)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(114)})
 
             finally:
 
@@ -2570,7 +2459,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2595,7 +2484,7 @@ class WebApi(Bottle):
                 if not FormData:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(105)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(105)})
 
                 elif idambiente != None and str(idambiente).isnumeric():
 
@@ -2651,13 +2540,13 @@ class WebApi(Bottle):
                                 if not Data:
                                 
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(641)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(641)})
                             
                             except:
 
                                 self.conn.rollback()                                
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(642)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(642)})
                             
                             finally:
                             
@@ -2666,22 +2555,22 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(106)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(106)})
                     
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(644)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(644)})
                     
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(645)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(645)})
             
             finally:
             
@@ -2690,7 +2579,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2746,13 +2635,13 @@ class WebApi(Bottle):
                                 if not Data:
 
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(541)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(541)})
                             
                             except:
 
                                 self.conn.rollback()                            
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(642)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(642)})
                             
                             finally:
                             
@@ -2761,7 +2650,7 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(644)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(644)})
 
                             for error in DataBeforeErrors:
 
@@ -2770,17 +2659,17 @@ class WebApi(Bottle):
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(651)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(651)})
                 
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(652)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(652)})
             
             finally:
             
@@ -2789,7 +2678,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2845,13 +2734,13 @@ class WebApi(Bottle):
                                 if not Data:
 
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(641)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(641)})
                             
                             except:
 
                                 self.conn.rollback()                            
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(661)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(661)})
                             
                             finally:
                             
@@ -2860,7 +2749,7 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(644)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(644)})
 
                             for error in DataBeforeErrors:
 
@@ -2869,17 +2758,17 @@ class WebApi(Bottle):
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(662)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(662)})
                 
                 else:
                 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
             
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(663)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(663)})
             
             finally:
             
@@ -2888,7 +2777,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -2964,13 +2853,13 @@ class WebApi(Bottle):
                                     if Data:
 
                                         Success = False
-                                        Errors.append({"msg": ErrorsDict.errorcode(671)})
+                                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(671)})
 
                                 except:
 
                                     self.conn.rollback()
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(672)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(672)})
 
                                 finally:
                                 
@@ -2980,7 +2869,7 @@ class WebApi(Bottle):
 
                                 self.conn.rollback()                
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(673)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(673)})
                             
                             finally:
 
@@ -2989,12 +2878,12 @@ class WebApi(Bottle):
                         else:
                         
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(674)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(674)})
                     
                     else:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(644)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(644)})
 
                         for error in DataBeforeErrors:
 
@@ -3003,12 +2892,12 @@ class WebApi(Bottle):
                 else:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(107)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(107)})
 
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(675)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(675)})
 
             finally:
 
@@ -3017,7 +2906,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -3054,7 +2943,7 @@ class WebApi(Bottle):
                 else:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(701)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(701)})
 
             else:
 
@@ -3065,7 +2954,7 @@ class WebApi(Bottle):
         except:
 
             Success = False
-            Errors.append({"msg": ErrorsDict.errorcode(702)})
+            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(702)})
             
 
         finally:
@@ -3136,13 +3025,13 @@ class WebApi(Bottle):
                             if not Data:
 
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(801)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(801)})
                         
                         except:
                             
                             self.conn.rollback()
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(802)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(802)})
                         
                         finally:
                         
@@ -3151,22 +3040,22 @@ class WebApi(Bottle):
                     except:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(803)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(803)})
 
                 elif MandatoryVarsExists == False:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(112)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(112)})
 
                 else: #MandatoryVarsTypes == False
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(113)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(113)})
             
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(114)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(114)})
 
             finally:
 
@@ -3175,7 +3064,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -3252,7 +3141,7 @@ class WebApi(Bottle):
                                 if not Data:
 
                                     Success = False
-                                    Errors.append({"msg": ErrorsDict.errorcode(811)})
+                                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(811)})
                             
                             except psycopg2.Error as ex:
 
@@ -3264,13 +3153,13 @@ class WebApi(Bottle):
                                 
                                 self.conn.rollback()
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(812) + psycopgError})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(812) + psycopgError})
 
                             except:
     
                                 self.conn.rollback()
                                 Success = False
-                                Errors.append({"msg": ErrorsDict.errorcode(812)})
+                                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(812)})
                             
                             finally:
                             
@@ -3278,27 +3167,27 @@ class WebApi(Bottle):
                         else:
                             
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(813)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(813)})
                     
                     except:
                     
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(814)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(814)})
 
                 elif MandatoryVarsExists == False:
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(112)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(112)})
 
                 else: #MandatoryVarsTypes == False
 
                     Success = False
-                    Errors.append({"msg": ErrorsDict.errorcode(113)})
+                    Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(113)})
             
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(114)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(114)})
 
             finally:
 
@@ -3307,7 +3196,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -3345,7 +3234,7 @@ class WebApi(Bottle):
 
                 self.conn.rollback()
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(601)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(601)})
             
             finally:
             
@@ -3355,7 +3244,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})            
 
@@ -3393,7 +3282,7 @@ class WebApi(Bottle):
 
                 self.conn.rollback()
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(401)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(401)})
 
             finally:
 
@@ -3404,7 +3293,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -3449,7 +3338,7 @@ class WebApi(Bottle):
 
                             self.conn.rollback()
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(411)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(411)})
 
                         finally:
 
@@ -3458,12 +3347,12 @@ class WebApi(Bottle):
                     elif str(idbusca) == "":
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
 
                     else:
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(109)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(109)})
 
                 else:
 
@@ -3476,7 +3365,7 @@ class WebApi(Bottle):
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(412)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(412)})
 
             finally:
 
@@ -3485,7 +3374,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -3523,7 +3412,7 @@ class WebApi(Bottle):
 
                 self.conn.rollback()
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(401)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(401)})
 
             finally:
 
@@ -3534,7 +3423,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -3572,7 +3461,7 @@ class WebApi(Bottle):
 
                 self.conn.rollback()
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(401)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(401)})
 
             finally:
 
@@ -3583,7 +3472,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -3629,7 +3518,7 @@ class WebApi(Bottle):
 
                             self.conn.rollback()
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(411)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(411)})
 
                         finally:
 
@@ -3638,12 +3527,12 @@ class WebApi(Bottle):
                     elif str(idbusca) == "":
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
 
                     else:
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(109)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(109)})
 
                 else:
 
@@ -3656,7 +3545,7 @@ class WebApi(Bottle):
             except:
 
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(412)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(412)})
 
             finally:
 
@@ -3665,7 +3554,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
 
@@ -3709,7 +3598,7 @@ class WebApi(Bottle):
                         else:
 
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(108)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(108)})
 
                             return({"success":Success,"errors":Errors,"data":Data})
 
@@ -3731,7 +3620,7 @@ class WebApi(Bottle):
             
                             self.conn.rollback()                            
                             Success = False
-                            Errors.append({"msg": ErrorsDict.errorcode(611)})
+                            Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(611)})
                         
                         finally:
                         
@@ -3740,7 +3629,7 @@ class WebApi(Bottle):
                     else:
 
                         Success = False
-                        Errors.append({"msg": ErrorsDict.errorcode(104)})
+                        Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(104)})
                 
                 else:
                 
@@ -3754,7 +3643,7 @@ class WebApi(Bottle):
             except:
             
                 Success = False
-                Errors.append({"msg": ErrorsDict.errorcode(621)})
+                Errors.append({"msg": ErrosFile.ErrorsDict.errorcode(621)})
             
             finally:
             
@@ -3763,7 +3652,7 @@ class WebApi(Bottle):
         else:
 
             # connectionErrors só será passado para usuarioid 1(suporte)
-            Errors = [{"msg": ErrorsDict.errorcode(300)}]
+            Errors = [{"msg": ErrosFile.ErrorsDict.errorcode(300)}]
 
             return json.dumps({"success": connectionStatus, "errors": Errors, "data": connectionData})
